select: fix int overflow of math.MaxInt64 on ARM (#6317)

+++ pkg/s3select/select_test.go
@@ -296,7 +296,7 @@ func TestMyParser(t *testing.T) {
 		if alias != table.alias {
 			t.Error()
 		}
-		if myLimit != table.myLimit {
+		if myLimit != int64(table.myLimit) {
 			t.Error()
 		}
 		if !reflect.DeepEqual(table.aggFuncs, aggFunctionNames) {
